basic neural mp qg diff policy exp working

+++ diffusion_policy/gym_util/sync_vector_env.py
@@ -44,7 +44,7 @@ class SyncVectorEnv(VectorEnv):
         self.observations = create_empty_array(
             self.single_observation_space, n=self.num_envs, fn=np.zeros
         )
-        self._rewards = np.zeros((self.num_envs,), dtype=np.float64)
+        self._rewards = np.zeros((self.num_envs,), dtype=np.float32)
         self._dones = np.zeros((self.num_envs,), dtype=np.bool_)
         # self._rewards = [0] * self.num_envs
         # self._dones = [False] * self.num_envs
